+++ react/ui-components/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.0.2-beta.11] - 2024-07-04
 ### New Changes
-- Added ActionBar,Header,ActionButton variant for Button
+- Added ActionBar, Header, ActionButton variant for Button
 - Updated Timeline Molecule 
 
 ## [0.0.2-beta.10] - 2024-06-28